turn on cosmological dimming in Sed.redshiftSED by default.
This required a change to the testCosmology.py unit test.  The
`absoluteGalaxyCatalog` class, which was supposed to return magnitudes
without a distance modulus, needed a getter which returned zero for
the cosmological distance modulus, since PhotometryGalaxies turns
cosmological dimming on and off based on the existence of methods to calculate
cosmological distnace modulus.