In testSed.py set warning filter to show all warnings
The default setting of the warnings module is to suppress identical
warnings that are raised from the same point in the code.  This was
causing a unittest to fail because the warning it was checking for
was being raised multiple times.  Setting this simplefilter disables
the warning suppression, allowing the unit test to pass.